Specify details of MCA Labeling framework use
When defining AudioChannel, SoundfieldGroup and GroupOfSoundfieldGroup
for use in the Multichannel Audio Labeling Framework (SMPTE ST 377-4) we
have to define: "MCA Tag Name" and "MCA Tag Symbol" for each of
them, and also which AudioChannels a SoundfieldGroup is comprised of.

Also introduces "Audio Description Studio Signal", which is used to
deliver Audio Description with control data in the UK and elsewhere.