Clarify order of Channels in 5.1 SoundfieldGroup
The table of MCA Labeling details now lists the constituent channels in
the same order as they are to be used in the Defualt Audio Layout. This
will minimise confusion

This commit also adds a clarification to the table of MCA Labeling
details, that the channels are presented in no particular order.